third_party/blink/web_tests/platform/generic/fast/forms/date-multiple-fields/date-multiple-fields-change-layout-by-value-expected.txt

Change multiple fields date input UI layout by value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS widthOfEmptyValue < widthOfOverflowValue is true
PASS widthOfEmptyValue is widthOfValidValue
PASS successfullyParsed is true

TEST COMPLETE
